requirements.txt
python-telegram-bot==20.3
Django==5.0.7
django-environ==0.11.2